build: migrate project toolchain to python 3.11 and ubuntu 22.04
This commit upgrades all of the project configuration files, GitHub Action workflows, and
Dockerfiles to use Python 3.11 and Ubuntu 22.04.

This commit also upgrades the pinned version of Tensorflow to 2.16.1 and PyTorch to 2.2.2, which
both have installable wheel support for Python 3.11. The detectron2 and pytorch-lightning packages
have been dropped due to lack of support and compatibility with these versions of PyTorch and
Python.