feat: add search functionality to v1 REST API
This feature adds search functionality to the REST API and integrates it
into the Users, Groups, and Queues endpoints.

The search functionality allows for multiple comma separated search terms
ANDed together. It supports multi character wildcards (*) and single
character wildcards (?). By default all searchable fields are searched. A
specific field can be specified with the field_name:search_term syntax.
Searchable fields are defined in the service layer of each resource.

If a query cannot be parsed, a SearchParseError will be raised, which is
handled by the REST API's error handler.