feat: implemented drafts for the v1 Queues endpoint
This feature implements drafts for the Queues endpoint. It includes the
drafts schema, controller, and service layers. Drafts are implemented in
a generic way that allows for the drafting functionality to be added to
other Resources with minimal code duplication.

Tests for drafting queues have been added to the queues test suite.